requirements.txt
fastapi==0.111.0
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.1
h11==0.14.0
httptools==0.6.1
idna==3.7
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
typing_extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
SQLAlchemy-Utils==0.41.2
alembic==1.13.1
asyncpg== 0.29.0
httpx==0.27.0
sqladmin==0.17.0
PyJWT==2.8.0
bcrypt==4.1.3
itsdangerous==2.2.0